Добавлены лайки, добавлено получение user_id через авторизованный аккаунт, добавлена загрузка файлов и прочее.

+++ src/main/java/org/app/breeze/entity/Comment.java
@@ -27,7 +27,7 @@ public class Comment {
     private String content;
 
     @Column(name = "likes_count")
-    private Integer likesCount = 0;
+    private Long likesCount = 0L;
 
     @Column(name = "created_at", updatable = false)
     @Temporal(TemporalType.TIMESTAMP)
